fail and try a few things
- Tried a method suggested by Julia to mark up/down macros whose
attributes are not in sysfs docs (these macros are found using
documented attrs). Not very successful. All the 106 macros (in driver
files) had a field which was in attributes.

- What I actually did was to look through macro names which did not have
the word 'attr' or 'attribute' in them (around 15) and manually ruled
out the ones which were not attribute declarers.

- Found a bug. To find the list of undocumented attrs, I was looking at
the symmteric difference of found_attrs and documented_attrs. This is
incorrect. Changed it to a set difference between found_attrs and
documented_attrs. (The order here matters). That brought down the number
from ~2000 to ~900. These are stored in result/undocumented_attrs.txt.

Signed-off-by: Aishwarya Pant <aishpant@gmail.com>